Meta & Tag Updates.
Updated Meta for <context.query> as it never returned a ListTag. it returns an element.
Added <context.query_map> to return a MapTag of the query. (Much more user friendly).
Minor meta grammatical updates.

Marked BlackCoyote's tutorial as out of date as the content was invalid, however left there for the examples.