Meta & Tag Updates. (#5)
* Meta & Tag Updates.

Updated Meta for <context.query> as it never returned a ListTag. it returns an element.
Added <context.query_map> to return a MapTag of the query. (Much more user friendly).
Minor meta grammatical updates.

Marked BlackCoyote's tutorial as out of date as the content was invalid, however left there for the examples.

* Further simplifification.

* Added URLDecoding to the query_map tag

* Moved the decoder to the correct place

* Had some sleep. Finally realised I'm an idiot.

* `, 2`